태그 카테고리
한 번이라도 참조된 태그 혹은 카테고리는 삭제가 불가능하도록 에러
메세지를 출력하도록 서버/클라이언트 로직 재설정.

1. 삭제 요청이 들어온 태그 혹은 카테고리를 찾고 _id 값을 가져오기
2. 모든 Blog 정보를 가져오고 오직 tags - name 정보만 populate 하기
3. dummy array를 생성하기
4. forEach 문을 이용해 모든 블로그에 사용된 태그 혹은 카테고리를 더미
배열에 할당하기
5. 1번에서 찾아둔 _id 값을 이용해 더미 배열에 includes 메소드를 이용해
true of false 여부 확인하기
6. 만약 true라면 블로그 생성에 사용된 태그 혹은 카테고리의 삭제가
불가하다는 에러 메세지 출력...

+++ backend/controllers/category/categoryCRUD.js
@@ -6,7 +6,7 @@ const { errorHandler } = require('../../helpers/dbErrorHandler');
 exports.create = async (req, res) => {
 	const { name } = req.body;
 	// new table --> slugify --> new-table
-	
+
 	let slug = slugify(name).toLowerCase();
 	let category = new Category({ name, slug });
 
@@ -47,9 +47,34 @@ exports.read = async (req, res, next) => {
 exports.remove = async (req, res) => {
 	const slug = req.params.slug.toLowerCase();
 	// {index: true} 한 이유, slug 위주로 queries 문 작성
+
+	// return res.status(404).json({message: "블로그 생성에 사용된 카테고리는 삭제할 수 없습니다..."})
+
 	try {
-		await Category.findOneAndRemove({ slug });
-		return res.status(200).json({ message: '카테고리가 성공적으로 삭제되었습니다.' });
+		let removedCategory = await Category.find({ slug }).select('_id');
+		let blogs = await Blog.find({})
+			.populate('categories', 'name')
+			.select(
+				'-photo -tags -like -title -body -excerpt -mtitle -mdesc -createdAt -updatedAt -_id -slug -postedBy -__v'
+			);
+
+		let usedCategoryList = [];
+
+		blogs.forEach((blog) => {
+			blog.categories.forEach((innerCategory) => {
+				usedCategoryList.push(innerCategory._id.toString());
+			});
+		});
+
+		let removedCategoryId = removedCategory[0]._id.toString();
+		let finder = usedCategoryList.includes(removedCategoryId);
+
+		if (!finder) {
+			await Category.findOneAndRemove({ slug });
+			return res.status(200).json({ message: '카테고리가 성공적으로 삭제되었습니다.' });
+		} else {
+			return res.status(400).json({ error: '블로그 생성에 사용된 카테고리는 삭제할 수 없습니다...' });
+		}
 	} catch (err) {
 		return res.status(400).json({ error: errorHandler(err) });
 	}
